Yet another take on the high level API
The cluster environment including Python/packages, files and environment
variables is now configured in the constructor. The ``run`` method on
the other hand takes an ``experiment_fn`` and the corresponding specs
for each of the task types.

Sidenote: turns out there is no need to copy the Kerberos ticket with
Skein. It requests a delegation token for the defaultFS on startup.