Removed ``max_restarts``
Implementing it correctly is harder than I anticipated initially. Once
one of the tasks fails, the cluster is usually (see [*]) paused until
the tasks is up again. However, re-acquiring the same port on container
restart is tricky as

* skein bumps the service instance ID on restart while the current
  implementation of ``_dispatch_task`` assumes it to be stable;
* the port could be reserved by a different TensorFlow application --
  sadly this will unblock (!) the training.

Therefore, the easiest/simple/bulletproof solution for now is to just
let the application crash, and let the user resubmit it to start training
from the latest checkpoint.

[*]: https://github.com/tensorflow/tensorflow/issues/7767#issuecomment-281841798